concat'd disparate firebase config files
